Move a bunch of test files around to start to organize them
