+++ test/utilities/DecentHatsModuleUtils.test.ts
@@ -23,7 +23,10 @@ import {
   KeyValuePairs__factory,
 } from '../../typechain-types';
 
-import { getGnosisSafeL2Singleton, getGnosisSafeProxyFactory } from '../GlobalSafeDeployments.test';
+import {
+  getGnosisSafeL2Singleton,
+  getGnosisSafeProxyFactory,
+} from '../global/GlobalSafeDeployments.test';
 
 import {
   getHatAccount,
